requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Flask==2.2.3
fastapi>=0.104.0
uvicorn>=0.24.0
APScheduler>=3.10.0
opencc-python-reimplemented>=0.1.7